pyproject.toml
[tool.poetry]
name = "fast-helm"
version = "0.1.0"
description = ""
authors = ["Mikalai Safranouski <mikalai.safranouski@slicelife.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
asyncpg = "^0.28.0"
sqlalchemy = {extras = ["async"], version = "^2.0.20"}
alembic = "^1.11.3"
python-decouple = "^3.8"
requests = "^2.31.0"
pytelegrambotapi = "^4.13.0"
aiohttp = "^3.8.5"
yfinance = "^0.2.28"


[tool.poetry.group.test.dependencies]
mypy = "^1.5.1"
black = "^23.7.0"
isort = "^5.12.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[mypy]
ignore_missing_imports = true
